+++ include/matrix.h
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 10:38:18 by hsarhan           #+#    #+#             */
-/*   Updated: 2022/12/21 17:49:18 by hsarhan          ###   ########.fr       */
+/*   Updated: 2022/12/23 12:03:13 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 # include "miniRT.h"
 
-typedef double			t_mat4[4][4];
-typedef double			t_mat3[3][3];
-typedef double			t_mat2[2][2];
+typedef float			t_mat4[4][4];
+typedef float			t_mat3[3][3];
+typedef float			t_mat2[2][2];
 
 typedef struct s_scene	t_scene;
 
@@ -26,11 +26,11 @@ void	mat_multiply(t_mat4 *res, const t_mat4 *m1, const t_mat4 *m2);
 void	mat_vec_multiply(t_vector *res, const t_mat4 *mat,
 			const t_vector *vec);
 void	transpose_matrix(t_mat4 *mat);
-void	translate_matrix(t_mat4 *mat, double x, double y, double z);
-void	scaling_matrix(t_mat4 *mat, double x, double y, double z);
+void	translate_matrix(t_mat4 *mat, float x, float y, float z);
+void	scaling_matrix(t_mat4 *mat, float x, float y, float z);
 
 // matrix_inverse.c
-double	det3(const t_mat3 *mat3);
+float	det3(const t_mat3 *mat3);
 void	submat4(t_mat3 *res, const t_mat4 *mat, int row, int col);
 void	mat_inverse(t_mat4 *res, const t_mat4 *mat);
 
@@ -38,10 +38,10 @@ void	mat_inverse(t_mat4 *res, const t_mat4 *mat);
 void	calculate_transforms(t_scene *scene);
 void	identity_matrix(t_mat4 *mat);
 
-void	rotation_matrix_x(t_mat4 *mat, double r);
-void	rotation_matrix_y(t_mat4 *mat, double r);
-void	rotation_matrix_z(t_mat4 *mat, double r);
-void	axis_angle(t_mat4 *rot_mat, const t_vector *ax, double angle);
+void	rotation_matrix_x(t_mat4 *mat, float r);
+void	rotation_matrix_y(t_mat4 *mat, float r);
+void	rotation_matrix_z(t_mat4 *mat, float r);
+void	axis_angle(t_mat4 *rot_mat, const t_vector *ax, float angle);
 
 // ! Wont be submitted
 void	print_mat4(const t_mat4 *mat);
