+++ include/miniRT.h
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 15:40:23 by hsarhan           #+#    #+#             */
-/*   Updated: 2022/12/21 22:14:27 by hsarhan          ###   ########.fr       */
+/*   Updated: 2022/12/23 12:03:13 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,6 +36,6 @@
 # define CAM_SPEED 0.5
 # define EPSILON 0.001
 # define TICK(X) clock_t X = clock()
-# define TOCK(X) printf("time %s: %g sec.\n", (#X), (double)(clock() - (X)) / CLOCKS_PER_SEC)
+# define TOCK(X) printf("time %s: %g sec.\n", (#X), (float)(clock() - (X)) / CLOCKS_PER_SEC)
 
 #endif
